requirements.txt
# RAG-based NLU-to-Tool Matching System
# Open-source dependencies

# Core ML/NLP
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
numpy>=1.24.3
scikit-learn>=1.3.0

# Data processing
pyyaml>=6.0.1
pandas>=2.0.3
python-dotenv>=1.0.0

# Logging and utilities
loguru>=0.7.2
rich>=13.4.2

# Web framework (for API)
fastapi>=0.100.0
uvicorn>=0.23.1

# MCP client support
httpx>=0.24.1
websockets>=11.0.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Development
black>=23.7.0
flake8>=6.0.0
